Jack - Peer Review
Had a look at the code, added some comments, all of them are tabbed to the right-hand side to make them easier to find. Mostly looks good, will speak to you more in person. 

Let me know if you're happy with this format for peer review or if you have any better ideas.