Finished the comments + fixed some issues with core functionality.
Fixed issue regarding to how the passenger was processed, before they were checked in but they were not added to the number of passengers on the flight until their baggage was added. So either the GUI would be required to lock them on the screen to add bags or as was decided

We went with a method that checks the details of the passenger first with a checkDetails method then processes the passenger by asking for the baggage details.